requirements.txt
anyio==4.11.0
certifi==2025.11.12
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.2.1
python-telegram-bot==22.5
rsa==4.9.1
sniffio==1.3.1
Telethon==1.42.0
typing_extensions==4.15.0
